Plot an extended changes-in-changes model

Plots the results of the ecic model, either along the percentiles or in an event-study fashion.

Usage

ecic_plot(
  object,
  es_type = c("aggregated", "for_quantiles", "for_periods"),
  perc_plot = NULL,
  periods_plot = NULL,
  xlab = NULL,
  ylab = "QTE \n",
  ylim = NULL,
  size = 2,
  zero_line = FALSE,
  legend_title = "Percentiles"
)

Arguments

object

An ecic_table object.

es_type

If an event study was estimated with ecic, you can choose the style of the ES plot. "aggregated" puts everything in one plot. "for_quantiles" generates one plot for each percentile. "for_periods" generates one plot for each period.

perc_plot

Which percentiles to plot.

periods_plot

Which periods to plot.

xlab

Alternative x-axis label

ylab

Alternative y-axis label.

ylim

Define the y-axis limits.

size

Size of the point estimates.

zero_line

Add a horizontal line at zero.

legend_title

Change the title of the legend.

Value

A ggplot2 object.